refactor(agent-status): split CodexToolStatusResolver into sectioned extensions
CodexToolStatusResolver.swift had grown to 1141 LOC, well past the repo's
~500 LOC guideline. Split verbatim along its existing MARK sections into
same-type extensions, all still @MainActor CodexToolStatusResolver:

- CodexToolStatusResolver.swift (63 LOC): struct, constants, init, and the
  suppression-window gates — the core.
- +Title.swift (392): volatile-title fast path, blocked-title ready
  clearing, needs-input/running title promotion, the running-status
  factory, and running-promotion diagnostics.
- +Idle.swift (218): ready-title needs-input recovery and idle-title ready
  surfacing.
- +ShellReturn.swift (125): stale-state clearing after a shell prompt
  returns, plus the shell/non-Codex-prompt predicates only it uses.
- +Predicates.swift (131): the handful of static status/title predicates
  used by both the running- and idle-title paths (readyTitleMayClearStatus,
  statusIsStalePlanModePrompt, etc.) — widened from private to internal
  since the split forces cross-file visibility; nothing else needed
  widening (helpers used by a single caller stayed colocated and private).
- +Interrupt.swift (272): user-input/progress promotion, the interrupt /
  suppression payload policy, and transient-state / transcript-context
  clearing.

Per-file Logger constants follow the existing WorklaneStore+*.swift
convention (each split file declares only the loggers it uses).

Ran xcodegen generate after adding the files.
